requirements.txt
openai>=1.40.0
dashscope>=1.20.0
httpx>=0.27.0
edge-tts>=6.1.0
faster-whisper>=1.0.0
ffmpeg-python>=0.2.0
playwright>=1.45.0
apscheduler>=3.10.0
celery>=5.4.0
redis>=5.0.0
sqlalchemy>=2.0.0
pyyaml>=6.0.0
feedparser>=6.0.0
click>=8.1.0
rich>=13.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
Pillow>=10.0.0
